refactor: remove SidebarHeader component and related styles for improved structure
