+++ packages/bruno-app/src/components/Sidebar/CloseWorkspace/index.js
