+++ packages/bruno-app/src/components/Sidebar/WorkspaceSelector/index.js
